+++ extensions/glyph-search/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Glyphs Changelog
 
+## [Fix Circled and Squared glyph clipping] - {PR_MERGE_DATE}
+
+- Fixed Circled and Squared glyphs being clipped in the grid.
+
 ## [Initial Version] - 2023-02-10
 
 Hello world!